image loading

+++ src/constants.rs
@@ -1,2 +1,2 @@
-pub const SCREEN_WIDTH: f64 = 336.0;
-pub const SCREEN_HEIGHT: f64 = 262.0;
+pub const SCREEN_WIDTH: f32 = 336.0;
+pub const SCREEN_HEIGHT: f32 = 262.0;
